backend/requirements.txt
fastapi[all]
uvicorn
pydantic
python-multipart
google-generativeai
python-dotenv
